+++ keanu-project/src/main/java/io/improbable/keanu/vertices/tensor/TensorVertex.java
@@ -42,17 +42,22 @@ public interface TensorVertex<T, TENSOR extends Tensor<T, TENSOR>, VERTEX extend
 
     @Override
     default VERTEX fillTriangular(boolean fillUpper, boolean fillLower) {
-        return null;
+        return wrap(new FillTriangularVertex<>(this, fillUpper, fillLower));
+    }
+
+    @Override
+    default VERTEX trianglePart(boolean upperPart) {
+        return wrap(new TrianglePartVertex<>(this, upperPart));
     }
 
     @Override
     default VERTEX triUpper(int k) {
-        return null;
+        return wrap(new TriUpperVertex<>(this, k));
     }
 
     @Override
     default VERTEX triLower(int k) {
-        return null;
+        return wrap(new TriLowerVertex<>(this, k));
     }
 
     @Override