+++ keanu-project/src/main/java/io/improbable/keanu/vertices/tensor/TriUpperVertex.java
@@ -3,9 +3,9 @@ package io.improbable.keanu.vertices.tensor;
 import com.google.common.base.Preconditions;
 import io.improbable.keanu.annotation.ExportVertexToPythonBindings;
 import io.improbable.keanu.tensor.Tensor;
-import io.improbable.keanu.tensor.TensorShape;
 import io.improbable.keanu.vertices.LoadVertexParam;
 import io.improbable.keanu.vertices.NonProbabilisticVertex;
+import io.improbable.keanu.vertices.SaveVertexParam;
 import io.improbable.keanu.vertices.Vertex;
 import io.improbable.keanu.vertices.tensor.number.floating.dbl.Differentiable;
 import io.improbable.keanu.vertices.tensor.number.floating.dbl.nonprobabilistic.diff.PartialDerivative;
@@ -13,32 +13,43 @@ import io.improbable.keanu.vertices.tensor.number.floating.dbl.nonprobabilistic.
 import java.util.Collections;
 import java.util.Map;
 
-public class DiagVertex<T, TENSOR extends Tensor<T, TENSOR>, VERTEX extends TensorVertex<T, TENSOR, VERTEX>>
+public class TriUpperVertex<T, TENSOR extends Tensor<T, TENSOR>, VERTEX extends TensorVertex<T, TENSOR, VERTEX>>
     extends UnaryTensorOpVertex<T, TENSOR, VERTEX> implements NonProbabilisticVertex<TENSOR, VERTEX>, Differentiable {
 
+    private static final String K_NAME = "k";
+
+    private final int k;
+
     @ExportVertexToPythonBindings
-    public DiagVertex(@LoadVertexParam(INPUT_NAME) TensorVertex<T, TENSOR, VERTEX> inputVertex) {
-        super(getDiagShape(inputVertex.getShape()), inputVertex, inputVertex.ofType());
+    public TriUpperVertex(@LoadVertexParam(INPUT_NAME) TensorVertex<T, TENSOR, VERTEX> inputVertex,
+                          @LoadVertexParam(K_NAME) int k) {
+        super(getTriShape(inputVertex.getShape()), inputVertex, inputVertex.ofType());
+        this.k = k;
     }
 
-    private static long[] getDiagShape(long[] inputShape) {
-        Preconditions.checkArgument(inputShape.length > 0, "Diag only operates on rank >= 1");
-        return TensorShape.getDiagResultShape(inputShape);
+    private static long[] getTriShape(long[] inputShape) {
+        Preconditions.checkArgument(inputShape.length >= 2, "Tri Upper only operates on rank >= 2");
+        return inputShape;
     }
 
     @Override
     protected TENSOR op(TENSOR input) {
-        return input.diag();
+        return input.triUpper(k);
     }
 
     @Override
     public PartialDerivative forwardModeAutoDifferentiation(Map<Vertex, PartialDerivative> derivativeOfParentsWithRespectToInput) {
         final PartialDerivative partial = derivativeOfParentsWithRespectToInput.get(inputVertex);
-        return PartialDerivative.createFromWrtOf(partial.getWrtOf(inputVertex.getRank()).diag(), this.getRank());
+        return PartialDerivative.createFromWrtOf(partial.getWrtOf(inputVertex.getRank()).triUpper(k), this.getRank());
     }
 
     @Override
     public Map<Vertex, PartialDerivative> reverseModeAutoDifferentiation(PartialDerivative partial) {
-        return Collections.singletonMap(inputVertex, new PartialDerivative(partial.get().diagPart()));
+        return Collections.singletonMap(inputVertex, new PartialDerivative(partial.get().triUpper(k)));
+    }
+
+    @SaveVertexParam(K_NAME)
+    public int getK() {
+        return k;
     }
 }