+++ keanu-project/src/main/java/io/improbable/keanu/vertices/tensor/number/operators/binary/MatrixMultiplicationVertex.java
@@ -9,6 +9,7 @@ import io.improbable.keanu.vertices.tensor.BinaryTensorOpVertex;
 import io.improbable.keanu.vertices.tensor.TensorVertex;
 import io.improbable.keanu.vertices.tensor.number.NumberTensorVertex;
 import io.improbable.keanu.vertices.tensor.number.floating.dbl.Differentiable;
+import io.improbable.keanu.vertices.tensor.number.floating.dbl.nonprobabilistic.diff.AutoDiffBroadcast;
 import io.improbable.keanu.vertices.tensor.number.floating.dbl.nonprobabilistic.diff.PartialDerivative;
 
 import java.util.HashMap;
@@ -41,22 +42,25 @@ public class MatrixMultiplicationVertex<T extends Number, TENSOR extends NumberT
     public Map<Vertex, PartialDerivative> reverseModeAutoDifferentiation(PartialDerivative derivativeOfOutputWithRespectToSelf) {
 
         PartialDerivative dOutputsWrtLeft = PartialDerivative
-            .matrixMultiplyAlongWrtDimensions(
+            .matrixMultiply(
                 derivativeOfOutputWithRespectToSelf,
                 right.getValue().toDouble(),
                 true
             );
 
         PartialDerivative dOutputsWrtRight = PartialDerivative
-            .matrixMultiplyAlongWrtDimensions(
+            .matrixMultiply(
                 derivativeOfOutputWithRespectToSelf,
                 left.getValue().toDouble(),
                 false
             );
 
+        int[] sumRight = AutoDiffBroadcast.dimensionsWithShapeChange(dOutputsWrtRight.get().getShape(), this.getRank(), right.getShape());
+        int[] sumLeft = AutoDiffBroadcast.dimensionsWithShapeChange(dOutputsWrtLeft.get().getShape(), this.getRank(), left.getShape());
+
         Map<Vertex, PartialDerivative> partials = new HashMap<>();
-        partials.put(left, dOutputsWrtLeft);
-        partials.put(right, dOutputsWrtRight);
+        partials.put(left, new PartialDerivative(dOutputsWrtLeft.get().sum(sumLeft)));
+        partials.put(right, new PartialDerivative(dOutputsWrtRight.get().sum(sumRight)));
 
         return partials;
     }
@@ -70,13 +74,17 @@ public class MatrixMultiplicationVertex<T extends Number, TENSOR extends NumberT
         PartialDerivative partialsFromLeft = PartialDerivative.matrixMultiplyAlongOfDimensions(
             dLeftWrtInput,
             right.getValue().toDouble(),
-            true
+            true,
+            left.getShape(),
+            this.getShape().length
         );
 
         PartialDerivative partialsFromRight = PartialDerivative.matrixMultiplyAlongOfDimensions(
             dRightWrtInput,
             left.getValue().toDouble(),
-            false
+            false,
+            right.getShape(),
+            this.getShape().length
         );
 
         return partialsFromLeft.add(partialsFromRight);