+++ keanu-project/src/test/java/io/improbable/keanu/vertices/tensor/number/floating/operators/unary/CholeskyDecompositionVertexTest.java
@@ -18,11 +18,11 @@ public class CholeskyDecompositionVertexTest {
     @Test
     public void canFindCholeskyDecomposition() {
         //Example from: https://en.wikipedia.org/wiki/Cholesky_decomposition
-        DoubleVertex A = new ConstantDoubleVertex(new double[]{
+        DoubleVertex A = new ConstantDoubleVertex(
             4, 12, -16,
             12, 37, -43,
             -16, -43, 98
-        }).reshape(3, 3);
+        ).reshape(3, 3);
 
         DoubleTensor expected = DoubleTensor.create(new double[]{
             2, 0, 0,